+++ src/main/utils/electron.utils.ts
@@ -9,3 +9,11 @@ export function showFilePickerSync(mainWindow: BrowserWindow, options: Omit<Open
   }
   return result;
 }
+export function showFolderPickerSync(mainWindow: BrowserWindow, options: Omit<OpenDialogOptions, 'properties' | 'securityScopedBookmarks'>): string | null {
+  let result = null;
+  if (mainWindow) {
+    const filename = dialog.showOpenDialogSync(mainWindow, {...options, properties: ['openDirectory']});
+    result = filename?.pop() || null;
+  }
+  return result;
+}